Adds validation that cluster methods remain undef
Remove cluster methods as elgible vectors

The purpose of a cluster method is that it really knows what "subvectors"
point to.  It is sensible that ...present_tense produce a data structure that
...present_tense_first_person_singular_number and
...present_tense_second_person_singular_number, etc. metaprogramatically
query.  Therefore this test ensures that while the long vectors respond_to?,
the cluster method is /not/ responable_to and requires /real/ definition
elsewhere.